abstract status
